pyproject.toml
[tool.poetry]
name = "service_metadata"
version = "0.1.0"
description = "This service is designed to create and update file metadata."
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.7"
common = "0.0.14"
logger = "0.1.0"
anyio = "3.4.0"
asyncpg = "0.25.0"
attrs = "21.2.0"
charset-normalizer = "2.0.9"
fastapi = "0.62.0"
FastAPI-SQLAlchemy = "0.2.1"
fastapi-utils = "0.2.1"
gunicorn = "20.0.4"
httptools = "0.1.1"
httpx = "0.18.2"
opentelemetry-api = "1.8.0"
opentelemetry-exporter-jaeger = "1.8.0"
opentelemetry-instrumentation = "0.27b0"
opentelemetry-instrumentation-asyncpg = "0.27b0"
opentelemetry-instrumentation-fastapi = "0.27b0"
opentelemetry-instrumentation-httpx = "0.27b0"
opentelemetry-sdk = "1.8.0"
protobuf = "3.19.1"
psycopg2-binary = "2.9.3"
pydantic = "1.8.2"
PyJWT = "1.4.2"
python-dotenv = "0.19.1"
python-json-logger = "0.1.11"
python-multipart = "0.0.5"
sqlalchemy = "1.4.28"
uvicorn = "0.12.3"
uvloop = "0.14.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.4"
pytest-env = "0.6.2"
coverage = "5.3"
flake8 = "4.0.1"
pytest-httpx = "0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pilot"
url = "https://git.indocresearch.org/api/v4/groups/pilot/-/packages/pypi/simple"
default = true
